board/requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
psycopg2==2.9.5
python-dotenv==0.21.0
